Drop SngContext; make /sng/manage self-contained
The shared context only existed to hand a picked .sng's bytes from the landing
to the editor route. Move the "Open .sng" action onto the editor page instead,
so /sng/manage owns its own state with plain useState and there's nothing to
share across routes. The landing is now purely navigational — both cards push to
/sng/manage — and the editor's empty state offers "Open an existing .sng"
(Modify) alongside the drop zone (Create). Removes SngContext, the shared
layout, and the two route-wrapper components.
