+++ spotify-clonehero-next/app/sng/components/SngEditor.tsx
@@ -1,18 +1,21 @@
 'use client';
 
-import {ArrowLeft, Download, FileArchive} from 'lucide-react';
+import {ArrowLeft, Download, FileArchive, FolderInput} from 'lucide-react';
 import {Button} from '@/components/ui/button';
 import type {FileEntry} from '@/components/chart-picker/chart-file-readers';
 import ChartInfoCard from './ChartInfoCard';
 import DropZone from './DropZone';
 import PackageFileTable from './PackageFileTable';
-import type {DownloadFormat} from '../SngContext';
+
+export type DownloadFormat = 'sng' | 'zip';
 
 interface SngEditorProps {
   files: FileEntry[];
   onAdd: (entries: FileEntry[]) => void;
   onDelete: (fileName: string) => void;
   onDownload: (format: DownloadFormat) => void;
+  /** Open an existing .sng, replacing the current package. */
+  onOpenSng: () => void;
   onBack: () => void;
 }
 
@@ -21,6 +24,7 @@ export default function SngEditor({
   onAdd,
   onDelete,
   onDownload,
+  onOpenSng,
   onBack,
 }: SngEditorProps) {
   const empty = files.length === 0;
@@ -56,6 +60,12 @@ export default function SngEditor({
 
         {/* Right: drop target + file list */}
         <div className="space-y-3">
+          {empty && (
+            <Button variant="secondary" className="w-full" onClick={onOpenSng}>
+              <FolderInput className="mr-2 h-4 w-4" />
+              Open an existing .sng
+            </Button>
+          )}
           <DropZone onAdd={onAdd} />
           <PackageFileTable files={files} onDelete={onDelete} />
         </div>
